Add visible item calculating logic.
The code is responsible for measuring estimated amount of cells that can
fit on the screen. This code can potentially reduce memory usage, as we
can keep a factor of visible cells on screen instead of guessing how
many could fit on screen.
The calculation is not dependant on any constant dimensions, therefore
it could potentially work for any view in the RecyclerView.